requirements.txt
fastapi
uvicorn
PyMuPDF
python-dotenv
mistralai